agreements: hide `util` from manage.py (#4290)
The agreements app has a django-admin command, `import_agreements`,
that uses a helper file, `util.py`, for S3 operations. Since `util.py`
resides in the commands directory, it gets improperly exposed
as a potential management command itself. It shows up in the first
section of available commands if you execute `./cfgov/manage.py help`.
If used, it will raise an AttributeError.

If we rename the file `_util.py`, it becomes a proper private utility.