move is_filter_selected jinja tag to v1 app
This change moves where the `is_filter_selected` Jinja function is
defined, from the sheerlike app to the v1 app. This function is used in
the filterable list rendering logic.

To test locally using a production dump, run a local server and visit
http://localhost:8000/about-us/blog/. Test and verify that filtering by
various options continues to work properly.

This change also removes the deprecated `event_filter_args` macro which
is not used anywhere in the codebase. It also removes the
`selected_filters_for_field` tag which is also not used.

This change is part of a larger goal to remove the deprecated sheerlike
app, initially by reducing dependencies on the `sheerlike.get_request`
method.