remove deprecated sheerlike indexer
This change removes the `sheer_index` management command and related
code in the `sheerlike` app, along with the related `cfgov/processors`
subdirectory, which is no longer needed.

This is part of a larger effort to completely remove the deprecated
sheerlike app entirely. Future work is needed to continue cleanup,
including removing related Djagno settings.