bump wagtail version to 1.13.2
This change increases the version of Wagtail used from 1.13.1 to 1.13.2.
This patch release fixes several bugs, including a set of 500 errors
caused by null characters in URLs.

To confirm this fix, visit http://localhost:8000/foo/?x=%00 against a
local server and note that it returns 404, not 500 as before.

Release notes for this version are at
http://docs.wagtail.io/en/v2.1.1/releases/1.13.2.html.

All unit tests continue to pass; as this is a patch release, there are
no major changes to functionality.