sheerlike cleanup: deprecate global_include tag
This change deprecates use of the sheerlike `global_include` tag that
allowed for inclusion of arbitrary Jinja2 templates in Django templates.
It replaces that tag with two specific tags that allow rendering of the
cf.gov site header and footer in Django templates. The header and footer
are themselves Jinja2 templates and so can't be included in Django
templates without custom code.

In `v1.templatetags.header_footer` there are now two new Django template
tags, `{% include_header %}` and `{% include_footer %}`, that render the
site header and footer in a Django template (after `{% load
header_footer %}` has been included in the template).

This change is intended to simplify the logic to include these modules.
I've moved the templates from their previous
`cfgov/jinja2/v1/_includes/on-demand` location to a more specific
`v1/jinja2/v1` folder. This is more inline with Django conventions; as
the Jinja2 engine is configured in settings with `'APP_DIRS': True`,
the `jinja2` subdirectory is the conventional place for app templates to
live.

Relatedly, I've removed the other remaining on-demand template,
`secondary-navigation.html`, as I couldn't find any code that currently
uses it:

https://github.com/search?q=user%3Acfpb+secondary-navigation.html&type=Code

To test locally, run a local server and visit
http://localhost:8000/know-before-you-owe/. This is a Django template
that includes the site header and footer, rendered using the Jinja2
template engine and these new tags.

See also new unit tests that validate behavior of these new template tags.

This change is part of a larger effort to remove the deprecated
sheerlike app entirely from cfgov-refresh.