Reduce Travis queue by moving docs deploy to it's own stage
The master branch doesn't need to run the test suites on ever PR merge,
but the docs deploy currently relies on the master branch. Additionally,
the docs should only deploy when a tag is released, otherwise the docs
don't represent what's on production. This PR moves the docs to it's own
stage that relies on a tagged release condition.