deprecate sheerlike.url_for Jinja2 function
As part of deprecation/cleanup of the sheerlike app, this change removes
the custom `url_for` template function in favor of the standard Django
`static` tag, which we pass into the Jinja2 environment.

To verify locally, visit:

- http://localhost:8000/owning-a-home/closing-disclosure/
- http://localhost:8000/owning-a-home/loan-estimate/

and verify that all images in the various steps load properly.

These images are all available to the static tag because we include the
`static_built` directory in `settings.STATICFILES_DIRS`:

https://github.com/cfpb/cfgov-refresh/blob/3e3aed43d3cef23701ed9d4aaae45c199c4454be/cfgov/cfgov/settings/base.py#L230

The source images are located in `unprocessed/img` and get copied into
`static_built` as part of the frontend build. Note that this does
require that the frontend build complete before these pages can be
viewed; but this is consistent with how other frontend assets are
generated.