Remove progressive enhancement from search query field
The results' counts next to the reg checkboxes in the sidebar need to
be updated when new results are pulled in. That process is a bit more
complex than just switching out the regs3k-results div with new HTML
and the easiest solution is to just use a standard POST when the search
query form is submitted.