remove deprecated Sheerlike query logic
This commit removes quite a bit of logic from the `sheerlike` app that
has been deprecated for some time, mostly around making queries to
Elasticsearch to retrieve Wordpress content.

Specifically, it removes these global Jinja2 variables: `queries`,
`more_like_this`, and `get_document`, as they are no longer used. It
removes use of the `sheerlike.finders.SheerlikeStaticFinder` staticfiles
finder, as there should be no more Sheer sites to find files from. And
it removes a sizeable number of classes and methods that are no longer
used anywhere.

It leaves behind a number of methods that are still used elsewhere in
the project and should/will be moved to a more appropriate place.