regs3k: Add search-term highlighting
This is a relatively simple option that makes search terms bold,
leaves a reasonable window of text around the search term for context
and thus isn't too jarring when jumping to the result's reference link.

Other options truncate the results right at the
first instance of the search term, which introduces drawbacks:
* Highlighting of the first instance of the term seems redundant, since
the terms are always the first first words of each result.
*  It often fails to provide reasonable context and can be disorienting
when clicking on the result link.