suppress database load output in refresh script
This change modifies the `refresh-data.sh` script so that the console
output from the database load step is suppressed via redirection to
`/dev/null`. The current process echos many, many lines of output when
loading a database dump which is not particularly useful.

This only redirects stdout, and so error messages should still get
displayed (if any occur).