bump retirement app to version 0.7.0
This PR bumps the version of the optional retirement app
(https://github.com/cfpb/retirement) from 0.6.2 to version 0.7.0. This
new version adds support for Django 1.11, among other things.

Because this new version isolates the retirement_api templates under a
new "retirement_api" Django namespace, this upgrade requires a small
change to urls.py to specify that namespace.

To test, verify that these URLs work when running locally:

- http://localhost:8000/consumer-tools/retirement/before-you-claim/

  Go through the entire flow, verify that the API calls work. Also
verify that the "Where do these numbers come from?" link works properly
to go to the About page.

- http://localhost:8000/consumer-tools/retirement/before-you-claim/es/

  Same as above, in Spanish.

(You'll need to either use a production dump or load the fixture from
the retirement repo for this to work.)

- http://localhost:8000/consumer-tools/retirement/before-you-claim/about/

- http://localhost:8000/consumer-tools/retirement/before-you-claim/about/es/