add new environment-based feature flag condition
This change deprecates the use of Wagtail Site-based feature flags,
instead using a new flag condition based on the value of
`settings.DEPLOY_ENVIRONMENT`.

Two new feature flag conditions are defined: "environment is" and
"environment is not". These can be set to match or not match against the
value of `DEPLOY_ENVIRONMENT`, as the user prefers.

We currently set the value of this setting by environment variable, and
it differs depending on the cf.gov environment. For example, this value
is "production" in production, "beta" for beta.cf.gov, etc.

This change unblocks the cleanup of our legacy second Wagtail Site for
content.cf.gov, which will simplify some workarounds that we have in the
code.

As a related change, this also moves the inclusion of `flag_enabled` and
`flag_disabled` in Jinja2 templates from the `sheerlike` app to the
`core` app. This is part of more general cleanup of the `sheerlike` app.