regs3k: Always deliver search result counts by regulation. (#4358)
Our search payload was stripping out result counts when filtering by
regulation. This makes sure the counts are always delivered.

Also, to avoid wasting queries, the page will skip haystack when
it's just serving the search-entry form for search terms. Once a
query string is provided, then it will process a full query and filter
by the source regulation as a starting point.

We're also taking this opportunity to drop the urls.py file, which
is not being used.