enforce use of single wagtail site
This change adds a new Django migration in the v1 app that deletes all
non-default Wagtail Sites from the database.

For historical reasons we have deployments that have more than one
Wagtail Site object: a main one and then a duplicate one ("content")
which was used in the past for sharing draft changes.

This second site is no longer needed, and so we want to standardize all
deployments so that they only have a single site.

Wagtail-flags dependencies on multiple sites were removed in #4361. See
also GHE#2375 for additional context on this change.

I've implemented this as a Django data migration to be able to most
easily apply it across all of our deployments and also locally for
developers. Alternatively we could implement this deletion manually, but
we'd have to do it server by server. I'm open to other suggestions about
how best to implement this in practice.

To test this, you can use the `refresh-data.sh` script to download a
production database. This new migration will be run automatically, and
delete the second site. If you run a local server and browse around,
there'll be several places where you see relative links where previously
they were absolute (this also fixes some minor bugs related to URLs,
such as GHE#2764).