regs3k: Clean up haystack query sequence.
The Haystack queries had some cruft, including defining the same query twice.

Unfortunately we can't get by with a single query, because we need to capture
counts for all regulations before narrowing down to the chosen regulation(s).
I've tried 'optimizing' by capturing the overall counts with a Counter object
and a single query, but bafflingly hitting Haystack for each count is faster.