regs3k: Lint paragraphs for italics and dashes. (#4364)
* regs3k: Lint paragraphs for italics and dashes.

QA testers identified two recurring format errors:
* Improper boldface on references such as **see** and **et seq.**
* Hyphens where em dashes should be.

This PR adds a paragraph linting routine to fix such cases.

Testing

You can check the linting by importing a fresh copy of Reg E:

If you have a copy of ECFR xml in develop-apps, this would import it:

```
./cfgov/manage.py runscript ecfr_importer --script-args 1005 develop-apps/ECFR-title12.xml
```

Then uncheck 'draft' status on the newly imported effective-version. You should be able to
check the following spots for linted formatting:

/1005/2/#(b)(2)
This spot shows both a bold 'see' and a bold 'et seq.'

Before linting:

After linting:

<http://localhost:8000/policy-compliance/rulemaking/regulations/1005/12/#c>

This spot should show a hyphen that is fixed to be an em dash.

Before linting:

After linting:

Note
Neither unicode nor html entities worked for passing an em dash character
from regdown to our templates, because of a combination of autoescaping
and the lack fo a regdown extension em dashes.

This PR adds a Markdown extension, `markdown-emdash` to enable this character.

An added benefit is that editors will have an easy way to add an em dash if
they need one. A triple hyphen – which is a convention used in other flavors
of Markdown – will translate to an em dash.

* Fix paragraph regex