remove sheerlike GlobalRequestMiddleware and app
This commit deprecates use of the
`sheerlike.middleware.GlobalRequestMiddleware`. This middleware did two
things:

1. Added a global `request` variable to the Jinja2 context, passed via
the local Python thread.

This is no longer needed as the request should be properly passed into
all Django and Jinja-rendered templates.

2. Adds the ability to access the request meta variables using
dictionary syntax, e.g. `request.headers['foo']` gets converted to
`request.META['HTTP_FOO']`.

We don't access headers this way anymore (`git grep headers -- *.py`).

All unit tests continue to pass.

This change was previously applied in #4334 but had to later be rolled
back in #4337 because of an incompatibility with the optional retirement
app and its dependence on a global `request` variable. This dependency
has since been removed (in #4357) and is no longer a blocker.

As this middleware was the last piece of the sheerlike app, this entire
app can now be removed.