Match deploy process used in ansible-consumer-finance
When the --fake-initial flag is present, django won't try to create
tables that already exist. This is necessary because some tables were
created outside of django, but we have since updated to track those
tables in django-migrations.

Now, the process we use in ./refresh-data will match the real deploy
process we use for cf.gov