regs3k: fix reference to more regs link
The changes in #4395 added a new `RegulationsList` block that included a
`more_regs_page` field that can be used to select a Wagtail page to link
to for a full set of regulations. The template for the block erroneously
refers to `more_regs_url`.

This commit fixes that error and properly provides a value for the page
for the existing unit tests. It also fixes the block template to
properly refer to block fields using `value.attribute`. See Wagtail
documentation here:

http://docs.wagtail.io/en/v2.1.1/topics/streamfield.html#template-rendering