Copy landing page for complaintdatabase from complaint satellite app
This code was originally in https://github.com/cfpb/complaint

The landing page (found at /data-research/consumer-complaints/) was the
only page that the complaint satellite app was still serving. With this
change, we can stop including complaint in our requirements and
deactivate the repository.

This update contains the view, templates, images, and static css
necessary to render the page.

Consolidate the two templates for the consumer complaint
database landing page into a single one.

Also includes the following changes:

1. Remove unused block `nemo_styles`.
2. Remove unused block `responsive_nav`.
3. Remove unused block `nemo_js`.
4. Remove unused block `app_js`.
5. Remove unused block `page_js`.
5. Remove unused template tag library `feature_flags`.
6. Moved S3 URLs from http://files.consumerfinance.gov to https://.

hardcode timely responses to 97%

The percent of timely responses to consumer complaints is hardcoded
elsewhere on cf.gov and so this change removes dynamic logic in this
code to instead use 97% directly.