make regulations3k a required django app
The `regulations3k` app is part of this repository and so is not
optional. Its dependency `treemodeladmin` is also not optional. This
change moves these apps into `settings.INSTALLED_APPS`.

In future it would be nice to more properly sort and order the list of
apps, but I have kept the scope of this PR as small as possible.