Regs3k: Use flag to disable URLs in production
This removes the settings condition from the REGULATIONS3K flag so that
it can be turned on by environment via the admin.

The plan is that we can use 'environment is not' and 'production' to allow
regs3k URLs to be active on dev servers and on Beta, for user testing and
for a pilot run of the new app. But the flagged URL will return 404
on production for any regs3k page.

Testing

You can test this locally by adding the 'environment is not' ... 'production'
condition to the REGULATIONS3K flag. This should allow the regs3k pages
to render locally (unless you have a local evironment variable set to
`DEPLOY_ENVIRONMENT=production`).