pin django-taggit for django version compatibility
Yesterday django-taggit released a new version (0.23.0) that is
incompatible with Django <1.11:

https://github.com/alex/django-taggit/blob/master/CHANGELOG.txt#L6

django-taggit is normally pulled in by Wagtail's setup.py, but
unfortunately in the version of Wagtail we are on (1.13.2) the setup.py
specifies `>=0.20,<1.0`, which pulls in this incompatible version:

https://github.com/wagtail/wagtail/blob/v1.13.2/setup.py#L26

This change pins django-taggit to its previous version (0.22.2) to
retain Django compatibility. This line should be removed once we upgrade
to Django 1.11.