remove get_protected_url from ItemIntroduction
This change removes the use of `get_protected_url` in the
ItemIntroduction block's template, and also properly handles the case
where a page with this block lacks a filterable ancestor.

The ItemIntroduction block has two places where it can potentially link
to a different filterable page that is its page's parent. In other
words, if an ItemIntroduction block is on a page like /foo/bar, and its
parent page at /foo is filterable, then there are two places where there
might be links to /foo in the block.

First, if the page has categories set, and the "Show category" checkbox
is checked in the ItemIntroduction, then the block will link to the
parent page like /foo?categories=something.

An example of a live page in production with this logic is
https://www.consumerfinance.gov/foia-requests/foia-electronic-reading-room/recommendation-memo-director/

Second, if the page has authors, then a "By Person's Name" element is
added when the block is rendered, and this element links to the parent
page like /foo?authors=persons-name.

An example of a live page in production with this logic is
https://www.consumerfinance.gov/about-us/blog/bureau-roundtables-seek-public-input/

This commit fixes the template logic to handle the case where the page
has no filterable parent. The previous logic would generate invalid
URLs in this case. From a process perspective, the ItemIntroduction
block should only live on a page that lives under a filterable parent.
From a technical perspective, however, there is nothing stopping someone
from creating a page that doesn't follow this pattern. This change adds
logic to handle this situation, even though it does not exist in
production.