remove custom relative page link handler
This commit removes our custom page link handler that makes it so that
all Wagtail page URLs inserted into rich text fields are always
relative.

It reverts behavior restored in #2904.

As of #4367 we no longer maintain two distinct Wagtail Sites, and so
this behavior is no longer needed.

Thanks to @richaagarwal for pointing out this legacy code.