properly pass template context to video player
This commit fixes an issue introduced by #4380 that introduced more
strict requirements on page/block templates using the "request" context
variable.

The video-player.html template requires this variable, but wasn't being
loaded with context. This change fixes that.

To test, use a production dump and visit

http://localhost:8000/practitioner-resources/library-resources/

as an example page using the video player.