Use focus state for skip to main content links
This adds some helper JavaScript that changes the behavior of the "skip
to main content" link. Currently there is a bug on mobile devices in
which the skip link causes the browser to visibly scroll to the main
content area, but the screen reader remains within the previous context.

This tiny JS module will set a tabindex of -1 and shift focus to the
<main> element, resolving this bug. The solution is based on one
recommended by this article: https://axesslab.com/skip-links/