Parse links middleware fix: Add back expand_db_html call (#4437)
## Overview
Fixes a bug introduced in https://github.com/cfpb/cfgov-refresh/pull/4363, which refactored `parse_links` and removed a call to `expand_db_html`.   As a result, I had added a call to the `richtext` filter [here](https://github.com/cfpb/cfgov-refresh/pull/4363/files#diff-6a8e044564accdfeb5d0bd4c115b59afL52) but as reported in GHE/CFGOV/platform/issues/2956, this is not comprehensive.  

## Notes
- Instead of adding several `richtext` calls across templates, which I believe would be hard to maintain for the same reasons our previous usage of `parse_links` was, I thought it would be best to add back in the call to `expand_db_html`.  
- We'll want to do some additional clean-up if we go with this approach, which is removing any template calls to `richtext`, since those would no longer be needed.  I'd rather do that clean-up work in a separate PR though, since it also involves updating some tests.

## Testing
This fixes the links reported in the GHE issue. To test, visit the pages reported in that issue and confirm the links are working on my branch. I also added a test to confirm that the internal page link representation gets expanded as expected.