support running unit tests against Django 1.11
Currently when trying to run Python unit tests against Django 1.11, for
example with this command:

tox -r -e unittest-py27-dj111-wag113-slow

the test runner command fails immediately, before any tests are run.
This change fixes a few things that are incompatible with Django 1.11,
which allows the test command to run, and individual test failures to be
reported. This does not actually make the tests work, but at least lets
them run.

All tests continue to pass with Django 1.8.

Specific changes are:

- As of 1.9, the use of settings.MIGRATION_MODULES to disable migrations
has changed, and so our logic in cfgov.settings.test_nomigrations must
as well. See Django docs at
https://docs.djangoproject.com/ko/2.1/releases/1.9/#migrations.
- As of 1.10, support for string view paths to url() has been removed,
and the module itself must be passed. See Django docs at
https://docs.djangoproject.com/ko/2.1/releases/1.10/#features-removed-in-1-10.
- The version of django-extensions we were using, 1.5.5, is not
compatible with Django 1.11 due to the existence of a migration that
tries to import from South. Support for Django 1.11 is added to
django-extensions in version 1.7.8. This PR bumps the version to the
latest one, 2.1.0, which will include Django 2.0 support in future.