use boto3 instead of boto for Django 1.11 compat
These set of changes change the AWS S3 package used by this project to
boto3 from boto. See this post discussing differences:

https://boto3.readthedocs.io/en/latest/guide/migration.html

Generally, boto3 is a more modern library which is better maintained.

Direct uses of boto have been replaced with boto3, and relatedly this
change modifies the django-storages backend to use its boto3 backend.
The django-storages dependency has been bumped in version from 1.1.8 to
1.5.0; unfortunately we cannot yet use the most recent version, see
requirements/libraries.txt for an explanation.

Some of the Django settings have been simplified or ordered
alphabetically related to this change. Specifically, we previously used
an `AWS_S3_ROOT` environment variable to set an `AWS_S3_ROOT` setting
that was used to instantiate a custom storages backend. This is no
longer necessary as the more recent versions of django-storages directly
support use of `AWS_LOCATION` to set the root path. As this is always
"f" on all systems, I've hardcoded it here.

These changes are necessary to add Django 1.11 compatibility (mainly
because django-storages 1.1.8 didn't support Django 1.11) but also get
us more up-to-date in our use of S3.