remove Django 1.11-incompatible form widget code
The conference registration form block currently makes use of some
Python code that customizes how its field widgets are rendered, to make
them compatible with Capital Framework.

Unfortunately, the hooks that this code (primarily in
cfgov/data_research/widgets.py) used are no longer compatible with
Django 1.11. As part of our Django 1.8 -> 1.11 upgrade, this PR removes
this customization, instead falling back on the standard widget
rendering for checkboxes and radio buttons (we can still keep our
customizations for text/email inputs and Textareas).

Some additional CSS styling is added to improve the look of these
buttons. Unfortunately this is no longer compatible with CF.

To test, create a BrowsePage in Wagtail and then add a Conference
Registration Form block to the Content field. Add some test data and
preview the page to see the buttons.