add django 1.11 support to custom field id render
Our `v1.forms.FilterableListForm` form class has a custom method
`render_with_id` that renders a form field with a custom `id` attribute.
This is used in templates like this:

{{ form.render_with_id(form.field_name, desired_field_id) }}

instead of using the default

{{ form.field_name }}

This adds an `id="desired_field_id"` when the form's `<input>` tag is
rendered.

The existing code is incompatible with Django 1.11 due to this change:

https://docs.djangoproject.com/en/1.11/ref/forms/widgets/#django.forms.Widget.render

In Django 1.11, the signature for the `Widget.render` method was changed
to add another parameter.

This commit modifies the code to be more generic and handle both method
signatures.

To verify that this still works, you can run a local server using a
production dump and visit http://localhost:8000/about-us/newsroom/.
Inspect the form elements in the filterable list (for example, the
Author input) and you'll see `id` elements like `id="authors"`.