remove duplicate definition of Django admin URLs
We erroneously include the Django admin URLs twice in our main URL file
(lines 444 and 461). On Django 1.11 this results in the following
warning:

?: (urls.W005) URL namespace 'admin' isn't unique. You may not be able
to reverse all URLs in this namespace

This commit removes the second definition. You can confirm that the
Django admin URLs still work locally by visiting
http://localhost:8000/django-admin/ with a local server (although note
this will require the change in #4474 to work properly).