Silence runtime warning in tests on dj111
Running unit tests on Django 1.11 produced the following warning:

.tox/unittest-py27-dj111-wag113-fast/lib/python2.7/site-packages/unidecode/__init__.py:47:
RuntimeWarning: Argument <type 'str'> is not an unicode object. Passing
an encoded string will likely have unexpected results.
  _warn_if_not_unicode(string)

The warning occurred when adding authors and tags (fields provided by
django-taggit) on instances of CFGOVPage in tests, e.g.
`page.authors.add('unencoded string')`.

To prevent this, we use `from __future__ import unicode_literals` where
possible. It was not possible in one test file that contains tests for
unicode conversion. So, in that file, we explicitly use unicode strings
in the tests that are affected by the warning.