Use parentNode instead of parentElement
Older versions of IE return null when selecting a parentElement that isn't
considered an element. This isn't a problem 99.9% of the time but IE
doesn't consider SVGs to be HTML elements so our `closest` method was
barfing when traversing SVG elements.

Our dom-traverse specs were written to expect `null` when `closest` is not
passed a selector or when the parent node wasn't found. To comply with
this, we wrap the traversal loop in a try/catch because parentNode throws
if a parent doesn't exist (while parentElement returns `null`).

https://stackoverflow.com/questions/8685739/difference-between-dom-parentnode-and-parentelement
https://developer.mozilla.org/en-US/docs/Web/API/Node/parentElement
https://developer.mozilla.org/en-US/docs/Web/API/Node/parentNode