fix django 1.11 compatibility of regs3k tests
Two unit tests introduced in #4506 aren't compatible with Django 1.11
due to a change in behavior in Django's redirect() shortcut introduced
in Django 1.9:

https://docs.djangoproject.com/en/2.1/releases/1.9/#http-redirects-no-longer-forced-to-absolute-uris

Prior to Django 1.9, calling redirect() with a relative URL would
automatically convert it to an absolute URL. The unit tests were
comparing the expected redirect path against an absolute URL, but that
no longer happens in Django 1.11.

I've changed the tests to instead use the Django assertRedirects helper:

https://docs.djangoproject.com/en/1.11/topics/testing/tools/#django.test.SimpleTestCase.assertRedirects

This works properly in both versions.