remove deprecated references to boxsizing.htc
The minified complaint CSS that was brought into cfgov-refresh in #4507
(and upated in #4515) includes references to a boxsizing.htc file. This
file was from a vendor package:

https://github.com/Schepp/box-sizing-polyfill

and was included in a frontend build task in the now-deprecated
complaint repo:

https://github.com/cfpb/complaint/blob/4d6293c771098318561921f70a0adce5bddc75b3/gulp/config.js#L92

Because this file doesn't exist in cfgov-refresh, references to it in
the CSS break use of Django collectstatic when run using
ManifestStaticFilesStorage, which validates all referenced files:

https://docs.djangoproject.com/en/2.1/ref/contrib/staticfiles/#manifeststaticfilesstorage

As this file is only needed to support older browsers, its inclusion can
be removed.

To see this diff with slightly nicer coloring, try `git diff
--color-words`.