Regs3k: Improve interpretation redirects
Interpretation redirects were not working for past versions of
appendix and section references after our launch-day hot-fixing.
This was not a pressing issue, because we have yet to turn past
versions on in production, but we hope to soon, and this PR restores
past version redirects for interpratations.

We also lost some test covereage in the hot-fix, and this restores test coverage.

Testing

Interp version redirects should work locally (if you have non-draft versions loaded)
or on the build5 server.

* /eregulations/1002-Interp/2016-16301
* /eregulations/1026-Appendices-Interp/2016-30730