add tests to custom makemessages command
This change adds a Python test for the v1 app's custom `makemessages`
Django management command, which overrides the default one documented at

https://docs.djangoproject.com/en/1.11/ref/django-admin/#django-admin-makemessages

Our custom makemessages command adds support for block tags like this:

{% trans %}
Multi-line
statement
here
{% endtrans %}

which is not supported in Django templates.

This change also removes support in this command for Django 1.8 now that
this project is solely targeted at Django 1.11.

To test this locally, you can verify that the unit tests passes, and
then remove our custom makemessages command (including the .pyc file if
on Python 2). The built-in command will fail as it does not detect the
multi-line Jinja2 tag.