update StdoutCapturingTestRunner to Django 1.11
This change updates the logic in our custom StdoutCapturingTestRunner
class to simplify it a bit and remove a reference to Django 1.8.

Unfortunately this change actually has the side effect of making the
unit tests for this custom test runner a bit more complicated.

Between 1.8 and 1.11, a change was made to address this Django ticket:

https://code.djangoproject.com/ticket/26981

The ability was added for Django TestRunner classes to provide custom
kwargs to the underlying unittest test runner, which is the thing that
actually runs the tests. This was implemented in this PR:

https://github.com/django/django/pull/7009

That PR added the ability for Django TestRunner classes to define a
custom get_test_runner_kwargs method to provide custom kwargs.

By default the unittest test runner (TextTestRunner) prints its status
to the console using stderr (this is the output that contains ...E..F
and shows how many tests are run). When we actually test our test
runner, we want to suppress this output.

See comment in the modified test file for a bit more explanation.