Upgrade django-storages to latest
 Upgrade django-storages library from 1.5.0 to its latest version
(1.7.1), which requires updating the `AWS_DEFAULT_ACL` setting.

`AWS_DEFAULT_ACL` determines what permissions (ACL, or Access Control List)
are added to individual files when they are uploaded to our S3 bucket.
That file's permissions are also dictated by the permissions of the
bucket that contains it.

In version 2.0, django-storages will change the default value of
`AWS_DEFAULT_ACL` from `public-read` to `None`, to improve security.
When set to `None`, it will not add the public-read ACL directly to each
file; intstead, files fall back on the bucket's ACL. We added
`AWS_DEFAULT_ACL = None` to silence a warning in 1.7.1, and to adopt the
newly preferred behavior early.

@chosak and I tested and found that this update will not result in a
change of behavior for us. Since the buckets in question have
`public-read` permissions set up, all files will be publicly readable.