update ETag behavior for Django 1.11
This change removes the definition of the now-deprecated Django setting
USE_ETAGS in favor of the ConditionalGetMiddleware, which creates ETags
in much the same way.

See this comment in the Django docs about settings.USE_ETAGS being
deprecated as of Django 1.11:

https://docs.djangoproject.com/en/1.11/ref/settings/#use-etags

To test generation of ETags, follow these steps after running a local
server:

1. Execute this curl command to retrieve the home page (this executes a
GET, dumps the headers to stdout, and suppresses the page output):

  curl -sSL -D - http://localhost:8000/ -o /dev/null

This should include some kind of ETag header in the response, like

  ETag: "a9a97c7a754b331982fdd9b8ea6293ef"

(this value may be different for you.) Note that the server responds
with an HTTP 200 OK status code.

2. Make another request, providing the ETag as a request header:

  curl -sSL -D - -H 'If-None-Match: "a9a97c7a754b331982fdd9b8ea6293ef"' http://localhost:8000/ -o /dev/null

(use your custom value if different.) The server should respond with an
HTTP 304 Not Modified, indicating that the page content has not changed.