update form widget rendering for Django 1.11
This change removes some customization that was needed to support custom
form widget rendering on the conference registration form in Django 1.8,
now that this is more built-in in Django 1.11.

We no longer need to define custom widgets in order to mark form inputs
as required or more easily pass input attributes.

This change only affects the rendering of the Wagtail block
ConferenceRegistrationForm, which can be added to the main content
section of a BrowsePage. Unfortunately, the easiest way to test this is
by manually creating a page of this type, adding this block, and hitting
Preview after entering the required fields.

If you inspect the rendered HTML, you'll see that required fields do in
fact have a "required" attribute and that text inputs have the desired
"a-text-input a-text-input__full" as specified by Capital Framework.