remove outdated section from installation docs
This change removes a troubleshooting section from the installation docs
which is a bit outdated.

First, now that we're on Django 1.11, it's been quite some time since
we've used South, so the instructions pertaining to it should no longer
be needed.

Second, the recommendation to globally uninstall all Python packages
(including virtualenv) seems overly broad, especially now that some
people are using the Docker workflow. If not using Docker, all work
should be done in a virtual environment, so the recommendation to
uninstall all packages feels too strong.

To test, `pip install -r requirements/manual.txt` then `mkdocs serve`,
and navigate to http://localhost:8000/installation/.